refactor: notification 관련 로직을 Header로 이동
