+++ frontend/src/context/UserContext.tsx
@@ -39,8 +39,6 @@ const UserProvider = ({
 }: PropsWithChildren<UserProvideProps>) => {
   const [isLoggedIn, setIsLoggedIn] = useState(initialData.isLoggedIn);
   const [memberId, setMemberId] = useState<number | null>(initialData.memberId);
-  const [notificationEventSource, setNotificationEventSource] =
-    useState<EventSource | null>(null);
 
   const { mutate: deleteRefreshToken } = useCreateLogout();
 
@@ -70,26 +68,6 @@ const UserProvider = ({
     setIsLoggedIn(false);
     setMemberId(null);
   };
-
-  const handleNotificationEventSource = (e: MessageEvent) => {
-    console.log(e, JSON.parse(e.data));
-  };
-
-  useEffect(() => {
-    if (memberId) {
-      const notification = new EventSource(
-        `${process.env.API_URL}/subscribe?id=${memberId}`
-      );
-      notification.addEventListener("sse", handleNotificationEventSource);
-      setNotificationEventSource(notification);
-      return;
-    } else {
-      notificationEventSource?.removeEventListener(
-        "sse",
-        handleNotificationEventSource
-      );
-    }
-  }, [memberId]);
 
   const value = { isLoggedIn, login, logout, memberId };
